requirements.txt
streamlitstreamlit
google-generativeai>=0.3.0